Communicator interface
Communicator replaces messenger as central means of communication.
Compared to messenger, communicator allows reuse through child
messengers, but guarantees global synchronisation and other important
properties for such messengers. Also contains a proof-of-concept
in-memory implementation and a set of tests that should be runnable
with any type of communicator (just subclass CommunicatorTest).